Eliminate TreeViewElement and ITreeView
