+++ src/TestCentric/testcentric.gui/Presenters/GroupDisplayStrategy.cs
@@ -40,8 +40,7 @@ namespace TestCentric.Gui.Presenters
         {
             base.OnTestFinished(result);
 
-            if (_grouping != null)
-                _grouping.OnTestFinished(result);
+            ApplyResultToGroup(result, false);
         }
 
         // TODO: Move this to TestGroup? Would need access to results.
@@ -65,6 +64,88 @@ namespace TestCentric.Gui.Presenters
             }
 
             return groupIndex;
+        }
+
+        public void Add(TreeNode treeNode)
+        {
+            _view.Add(treeNode);
+        }
+
+        public void ApplyResultToGroup(ResultNode result, bool updateImages)
+        {
+            var treeNodes = GetTreeNodesForTest(result);
+
+            // Result may be for a TestNode not shown in the tree
+            if (treeNodes.Count == 0)
+                return;
+
+            // This implementation ignores any but the first node
+            // since changing of groups is currently only needed
+            // for groupings that display each node once.
+            var treeNode = treeNodes[0];
+            var oldParent = treeNode.Parent;
+            var oldGroup = oldParent.Tag as TestGroup;
+
+            // We only have to proceed for tests that are direct
+            // descendants of a group node.
+            if (oldGroup == null)
+                return;
+
+            var newGroup = _grouping.SelectGroups(result)[0];
+
+            // If the group didn't change, we can get out of here
+            if (oldGroup == newGroup)
+                return;
+
+            var newParent = newGroup.TreeNode;
+
+            _view.InvokeIfRequired(() =>
+            {
+                oldGroup.RemoveId(result.Id);
+                // TODO: Insert in order
+                newGroup.Add(result);
+
+                // Remove test from tree
+                treeNode.Remove();
+
+                // If it was last test in group, remove group
+                if (oldGroup.Count == 0)
+                    oldParent.Remove();
+                else // update old group
+                {
+                    oldParent.Text = GroupDisplayName(oldGroup);
+                    if (updateImages)
+                        oldParent.ImageIndex = oldParent.SelectedImageIndex = oldGroup.ImageIndex =
+                            CalcImageIndexForGroup(oldGroup);
+                }
+
+                newParent.Nodes.Add(treeNode);
+                newParent.Text = GroupDisplayName(newGroup);
+                newParent.Expand();
+
+                if (updateImages)
+                {
+                    var imageIndex = DisplayStrategy.CalcImageIndex(result.Outcome);
+                    if (imageIndex >= TestTreeView.SuccessIndex && imageIndex > newGroup.ImageIndex)
+                        newParent.ImageIndex = newParent.SelectedImageIndex = newGroup.ImageIndex = imageIndex;
+                }
+
+                if (newGroup.Count == 1)
+                {
+                    _view.Clear();
+                    TreeNode topNode = null;
+                    foreach (var group in _grouping.Groups)
+                        if (group.Count > 0)
+                        {
+                            Add(group.TreeNode);
+                            if (topNode == null)
+                                topNode = group.TreeNode;
+                        }
+
+                    if (topNode != null)
+                        topNode.EnsureVisible();
+                }
+            });
         }
 
         #endregion
@@ -114,7 +195,7 @@ namespace TestCentric.Gui.Presenters
                     treeNode.Expand();
                     if (group.Count > 0)
                     {
-                        _view.Tree.Add(treeNode);
+                        _view.Add(treeNode);
                         if (topNode == null)
                             topNode = treeNode;
                     }
