+++ src/TestCentric/tests/Presenters/TestTree/WhenTestCaseCompletes.cs
@@ -45,7 +45,7 @@ namespace TestCentric.Gui.Presenters.TestTree
             _model.Events.TestLoaded += Raise.Event<TestNodeEventHandler>(new TestNodeEventArgs(testNode));
             _model.Events.TestFinished += Raise.Event<TestResultEventHandler>(new TestResultEventArgs(resultNode));
 
-            _view.Tree.Received().SetImageIndex(Arg.Compat.Any<TreeNode>(), expectedIndex);
+            _view.Received().SetImageIndex(Arg.Compat.Any<TreeNode>(), expectedIndex);
         }
     }
 }
