+++ src/TestCentric/tests/Presenters/TestTree/WhenTestSuiteCompletes.cs
@@ -43,7 +43,7 @@ namespace TestCentric.Gui.Presenters.TestTree
             _model.Events.TestLoaded += Raise.Event<TestNodeEventHandler>(new TestNodeEventArgs(testNode));
             _model.Events.SuiteFinished += Raise.Event<TestResultEventHandler>(new TestResultEventArgs(resultNode));
 
-            _view.Tree.Received().SetImageIndex(Arg.Compat.Any<TreeNode>(), expectedIndex);
+            _view.Received().SetImageIndex(Arg.Compat.Any<TreeNode>(), expectedIndex);
         }
     }
 }
